requirements.txt
openenv-core
pydantic>=2.0.0
google-genai
python-dotenv
pytest>=7.0.0
groq
sentence-transformers
fastapi
uvicorn